fix slotmapping dtype

+++ vllm_ascend/worker/v2/model_runner.py
@@ -268,9 +268,7 @@ class NPUModelRunner(GPUModelRunner):
             num_computed_tokens_cpu=self.req_states.
             num_computed_tokens_cpu[idx_mapping_cpu],
             block_tables=block_tables,
-            # torch_npu._reshape_and_cache operator requires slot_mappings to
-            # be torch.int32.
-            slot_mappings=slot_mappings.to(torch.int32),
+            slot_mappings=slot_mappings,
             kv_cache_config=self.kv_cache_config,
             attn_state=attn_state,
         )
